change data type

+++ day16.cpp
@@ -1,7 +1,7 @@
 #include<bits/stdc++.h>
 using namespace std;
 
-int gcd(int a,int b){
+long long int gcd(long long int a,long long int b){
     while(b!=0){ //loop till b not 0
         int rem=a%b;
         a=b;
@@ -10,11 +10,11 @@ int gcd(int a,int b){
     return a; //as a store ans
 }
 
-int lcm(int a,int b){
+long long int lcm(long long int a,long long int b){
     return (a*b)/gcd(a,b); // as a*b = lcm(a,b) * gcd(a,b)
 }
 int main(){
-    int a,b;
+    long long int a,b;
     cin>>a>>b;
     cout<<lcm(a,b)<<endl;
     return 0;
